grabscreen.lua: tap and hold for >=500ms emulates click-drag instead of mouse wheel.
As a user, one wants to avoid xochitl interpreting a left or right swipe (previous
or next page) as that would draw over the painted Pi screen.
(Since commit "swiping from below (...)" this is dramatic any more, but anyway...)
The rM's swipe gesture appears to be time-sensitive in respect to the *speed* with
which the swipe is made -- only a relatively large velocity of the finger on the
surface leads to changing the page.
In any case, to be on the safe side, in case one wants to drag a horizontal line,
it is always possible to make the route curved to some degree in order for rM to
not interpret it. For us, only the start and end position matters.