grabscreen.lua: on start, clear the upper area on the rM (except left/right borders).
The left border must not be cleared because of the "eye". The right one may contain
the circled cross.

Move the "locked up" marker (which diagnoses a condition that is believed to be
fully fixed, but kept for the time being) to the left so that it is also cleared.