grabscreen.lua: issue rM refresh request as bounding box of all updated rects.
This is simpler than the previous two-staged rect merge logic and after
using it for a small while, I *believe* it to also be slightly better in
terms of not leaving around as much "garbage" (outdated pixels) on the rM
screen. (One example where outdated pixels can be observed is when the character
cursor in a text editor is moved many times in quick succession.)
To be concrete, in Emacs, after invoking iswitchb-buffer and typing a prefix,
the result appears legible, whereas previously it was often garbled.