Add make_layouts_as_single_lua.sh, generating 'layouts/.lua'
from previously generated '.codepoints' and layout files (e.g. 'us.basic').

This seems like the easier way to embed these data into Lua via require().
The alternative would have been somehow modifying / providing an own
require() so that it does not replace '.' with the directory separator,
which would
 - necessitate reimplementing require() functionality that is not merely
   just the part we want changed, and
 - also entail mkapp.lua to be kept in sync with those changes.

A start was attempted, and this is what the bump of the 'ljclang'
submodule is about.

Of course, it is reasonable to say that we might as well from the start
choose file names which are permissible to / convenient for require().